ml/requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
psycopg2-binary>=2.9
python-dotenv>=1.0
# plotting is optional — keep off for headless Railway / CI runs
matplotlib>=3.7
